requirements.txt
Flask
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
Werkzeug
torch
torchvision
Pillow
XlsxWriter